skills/hughpyle/keep/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "keep-skill"
version = "0.8.1"
description = "Reflective memory - remember and search documents by meaning"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = {text = "MIT"}
authors = [{name = "Hugh Pyle"}]
keywords = ["semantic-memory", "vector-search", "embeddings", "chromadb", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "chromadb>=0.4",
    "typer>=0.9",
    "tomli-w>=1.0",
]

[project.urls]
Homepage = "https://github.com/hughpyle/keep"
Repository = "https://github.com/hughpyle/keep"

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.40.0"]
gemini = ["google-genai>=1.0.0"]
sentence-transformers = ["sentence-transformers>=2.2"]
documents = [
    # Document text extraction (PDF, HTML)
    "pypdf>=5.0",
    "beautifulsoup4>=4.9",
]
mlx = [
    "mlx>=0.10",
    "mlx-lm>=0.10",
]
local = [
    # Apple Silicon: MLX for embedding/summarization
    "mlx>=0.10; platform_system == 'Darwin' and platform_machine == 'arm64'",
    "mlx-lm>=0.10; platform_system == 'Darwin' and platform_machine == 'arm64'",
    # Fallback: sentence-transformers for embedding
    "sentence-transformers>=2.2",
    # Document extraction
    "pypdf>=5.0",
    "beautifulsoup4>=4.9",
]
openclaw = [
    # For OpenClaw integration: use configured models
    "anthropic>=0.18.0",
    "google-genai>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
keep = "keep.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["keep"]

[tool.hatch.build.targets.sdist]
include = [
    "/keep",
    "/SKILL.md",
    "/README.md",
    "/LICENSE",
]
